# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: FETAL-GUARD API
  description: |
    Backend API for FETAL-GUARD: Wearable Early Warning System for fetal monitoring.
    Provides endpoints for patient management, device pairing, monitoring sessions,
    alerts, and clinician dashboard.
  version: 1.0.0
  contact:
    name: FETAL-GUARD Team
    email: support@fetal-guard.id
  license:
    name: Proprietary
servers:
  - url: https://api.fetal-guard.id/v1
    description: Production server
  - url: https://staging-api.fetal-guard.id/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: List all patients (clinician only)
      parameters:
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /patients/{patient_id}:
    get:
      tags: [Patients]
      summary: Get patient details
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      tags: [Patients]
      summary: Update patient profile
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
...
  /devices/{device_id}:
    get:
      tags: [Devices]
      summary: Get device details
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'

    delete:
      tags: [Devices]
      summary: Unpair device
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Device unpaired
...
  /clinician/patients:
    get:
      tags: [Clinician]
      summary: Get clinician's patient list
      parameters:
        - name: risk_filter
          in: query
          schema:
            type: string
            enum: [all, high, active_alert]
      responses:
        '200':
          description: Patient list for clinician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClinicianPatientView'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PatientId:
      name: patient_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        patient_id:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        birth_date:
          type: string
          format: date
        lmp_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        gestational_age_weeks:
          type: integer
        gestational_age_days:
          type: integer
        weight_kg:
          type: number
        height_cm:
          type: number
        bmi:
          type: number
        medical_history:
          type: object
          properties:
            diabetes:
              type: boolean
            hypertension:
              type: boolean
            heart_disease:
              type: boolean
            preeclampsia:
              type: boolean
            other:
              type: string
        medications_current:
          type: array
          items:
            type: string
        emergency_contact:
          type: object
          properties:
            name:
              type: string
            relation:
              type: string
            phone:
              type: string
        clinic_id:
          type: string
        clinician_id:
          type: string
        consent_timestamp:
          type: string
          format: date-time
        consent_version:
          type: string
        privacy_flags:
          type: object
          properties:
            location_share_allowed:
              type: boolean
            upload_allowed_wifi_only:
              type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    PatientSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        gestational_age_weeks:
          type: integer
        last_session_date:
          type: string
          format: date-time
        current_risk_level:
          type: string
          enum: [low, medium, high]
        active_alerts:
          type: integer
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
        weight_kg:
          type: number
        medications_current:
          type: array
          items:
            type: string
        emergency_contact:
          type: object
        privacy_flags:
          type: object
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_serial:
          type: string
        device_model:
          type: string
        device_firmware_version:
          type: string
        patient_id:
          type: string
        paired_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        battery_level:
          type: integer
        status:
          type: string
          enum: [active, disconnected, low_battery]
...
    ClinicianPatientView:
      type: object
      properties:
        patient_id:
          type: string
        name:
          type: string
        gestational_age_weeks:
          type: integer
        last_session:
          type: string
          format: date-time
        current_risk:
          type: string
        active_alerts:
          type: integer
        last_fhr:
          type: number
...